Module Storage_functors.Make_indexed_carbonated_data_storage

Parameters

C : Tezos_raw_protocol_009_PsFLoren.Raw_context.T

I : INDEX

V : Tezos_raw_protocol_009_PsFLoren.Storage_sigs.VALUE

Signature

type t = C.t

type context = t

type key = I.t

An abstract type for keys

type value = V.t

The type of values

val mem : context -> key -> (Tezos_raw_protocol_009_PsFLoren.Raw_context.t * bool) Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

Tells if a given key is already bound to a storage bucket. Consumes Gas_repr.read_bytes_cost Z.zero.

val get : context -> key -> (Tezos_raw_protocol_009_PsFLoren.Raw_context.t * value) Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

Retrieve a value from the storage bucket at a given key ; returns Storage_error Missing_key if the key is not set ; returns Storage_error Corrupted_data if the deserialisation fails. Consumes Gas_repr.read_bytes_cost <size of the value>.

val find : context -> key -> (Tezos_raw_protocol_009_PsFLoren.Raw_context.t * value option) Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

Retrieve a value from the storage bucket at a given key ; returns None if the value is not set ; returns Storage_error Corrupted_data if the deserialisation fails. Consumes Gas_repr.read_bytes_cost <size of the value> if present or Gas_repr.read_bytes_cost Z.zero.

val update : context -> key -> value -> (Tezos_raw_protocol_009_PsFLoren.Raw_context.t * int) Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

Updates the content of a bucket ; returns A Storage_Error Missing_key if the value does not exists. Consumes serialization cost. Consumes Gas_repr.write_bytes_cost <size of the new value>. Returns the difference from the old to the new size.

val init : context -> key -> value -> (Tezos_raw_protocol_009_PsFLoren.Raw_context.t * int) Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

Allocates a storage bucket at the given key and initializes it ; returns a Storage_error Existing_key if the bucket exists. Consumes serialization cost. Consumes Gas_repr.write_bytes_cost <size of the value>. Returns the size.

val add : context -> key -> value -> (Tezos_raw_protocol_009_PsFLoren.Raw_context.t * int * bool) Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

Allocates a storage bucket at the given key and initializes it with a value ; just updates it if the bucket exists. Consumes serialization cost. Consumes Gas_repr.write_bytes_cost <size of the new value>. Returns the difference from the old (maybe 0) to the new size, and a boolean indicating if a value was already associated to this key.

val add_or_remove : context -> key -> value option -> (Tezos_raw_protocol_009_PsFLoren.Raw_context.t * int * bool) Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

When the value is Some v, allocates the data and initializes it with v ; just updates it if the bucket exists. When the value is None, delete the storage bucket when the value ; does nothing if the bucket does not exists. Consumes serialization cost. Consumes the same gas cost as either remove or init_set. Returns the difference from the old (maybe 0) to the new size, and a boolean indicating if a value was already associated to this key.

val remove_existing : context -> key -> (Tezos_raw_protocol_009_PsFLoren.Raw_context.t * int) Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

Delete a storage bucket and its contents ; returns a Storage_error Missing_key if the bucket does not exists. Consumes Gas_repr.write_bytes_cost Z.zero. Returns the freed size.

val remove : context -> key -> (Tezos_raw_protocol_009_PsFLoren.Raw_context.t * int * bool) Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

Removes a storage bucket and its contents ; does nothing if the bucket does not exists. Consumes Gas_repr.write_bytes_cost Z.zero. Returns the freed size, and a boolean indicating if a value was already associated to this key.