Module Tezos_raw_protocol_009_PsFLoren.Receipt_repr

type balance =

| Contract of Contract_repr.t

| Rewards of Tezos_crypto.Signature.public_key_hash * Cycle_repr.t

| Fees of Tezos_crypto.Signature.public_key_hash * Cycle_repr.t

| Deposits of Tezos_crypto.Signature.public_key_hash * Cycle_repr.t

Places where tezzies can be found in the ledger's state.

type balance_update =

| Debited of Tez_repr.t

| Credited of Tez_repr.t

A credit or debit of tezzies to a balance.

type update_origin =

| Block_application	(*
Update from a block application
*)

| Protocol_migration	(*
Update from a protocol migration
*)

An origin of a balance update

type balance_updates = (balance * balance_update * update_origin) list

A list of balance updates. Duplicates may happen.

val balance_updates_encoding : balance_updates Tezos_base__TzPervasives.Data_encoding.t

val cleanup_balance_updates : balance_updates -> balance_updates

Remove zero-valued balances from a list of updates.