Module Tezos_raw_protocol_009_PsFLoren.Tez_repr

type t

type tez = t

val zero : t

val one_mutez : t

val one_cent : t

val fifty_cents : t

val one : t

val (-?) : t -> t -> t Tezos_protocol_environment_009_PsFLoren.Error_monad.tzresult

val (+?) : t -> t -> t Tezos_protocol_environment_009_PsFLoren.Error_monad.tzresult

val (*?) : t -> int64 -> t Tezos_protocol_environment_009_PsFLoren.Error_monad.tzresult

val (/?) : t -> int64 -> t Tezos_protocol_environment_009_PsFLoren.Error_monad.tzresult

val to_mutez : t -> int64

val of_mutez : int64 -> t option

of_mutez n (micro tez) is None if n is negative

val of_mutez_exn : int64 -> t

of_mutez_exn n fails if n is negative. It should only be used at toplevel for constants.

val mul_exn : t -> int -> t

It should only be used at toplevel for constants.

val encoding : t Tezos_protocol_environment_009_PsFLoren.Data_encoding.t

val to_int64 : t -> int64

include Tezos_protocol_environment_009_PsFLoren.Compare.S with type t := t

val (=) : t -> t -> bool

val (<>) : t -> t -> bool

val (<) : t -> t -> bool

val (<=) : t -> t -> bool

val (>=) : t -> t -> bool

val (>) : t -> t -> bool

val compare : t -> t -> int

val equal : t -> t -> bool

val max : t -> t -> t

val min : t -> t -> t

val pp : Tezos_protocol_environment_009_PsFLoren.Format.formatter -> t -> unit

val of_string : string -> t option

val to_string : t -> string