Module Tezos_raw_protocol_009_PsFLoren.Voting_period_storage

val init : Raw_context.t -> Voting_period_repr.t -> (Raw_context.t, Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.error Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.trace) Stdlib.result Lwt.t

val init_first_period : Raw_context.t -> start_position:int32 -> (Raw_context.t, Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.error Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.trace) Stdlib.result Lwt.t

Sets the initial period to {voting_period = root; kind = Proposal; start_position}.

val reset : Raw_context.t -> (Raw_context.t, Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.error Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.trace) Stdlib.result Lwt.t

Increment the index by one and set the kind to Proposal.

val succ : Raw_context.t -> (Raw_context.t, Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.error Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.trace) Stdlib.result Lwt.t

Increment the index by one and set the kind to its successor.

val get_current : Raw_context.t -> (Voting_period_repr.t, Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.error Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.trace) Stdlib.result Lwt.t

val get_current_kind : Raw_context.t -> (Voting_period_repr.kind, Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.error Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.trace) Stdlib.result Lwt.t

val is_last_block : Raw_context.t -> (bool, Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.error Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.trace) Stdlib.result Lwt.t

Returns true if the context level is the last of current voting period.

val get_current_info : Raw_context.t -> (Voting_period_repr.info, Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.error Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.trace) Stdlib.result Lwt.t

val get_rpc_fixed_current_info : Raw_context.t -> (Voting_period_repr.info, Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.error Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.trace) Stdlib.result Lwt.t

val get_rpc_fixed_succ_info : Raw_context.t -> (Voting_period_repr.info, Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.error Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.trace) Stdlib.result Lwt.t