Module Tezos_raw_protocol_009_PsFLoren__Raw_context

Errors

type Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.error += Too_many_internal_operations

type missing_key_kind =

| Get

| Set

| Del

| Copy

type storage_error =

| Incompatible_protocol_version of string

| Missing_key of string list * missing_key_kind

| Existing_key of string list

| Corrupted_data of string list

An internal storage error that should not happen

type Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.error += Storage_error of storage_error

type Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.error += Failed_to_parse_parameter of bytes

type Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.error += Failed_to_decode_parameter of Tezos_protocol_environment_009_PsFLoren__Environment.Data_encoding.json * string

val storage_error : storage_error -> 'a Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult

Abstract Context

type t

Abstract view of the context. Includes a handle to the functional key-value database (Context.t) along with some in-memory values (gas, etc.).

type root = t

val prepare : level:Tezos_protocol_environment_009_PsFLoren__Environment.Int32.t -> predecessor_timestamp:Tezos_protocol_environment_009_PsFLoren__Environment.Time.t -> timestamp:Tezos_protocol_environment_009_PsFLoren__Environment.Time.t -> fitness:Tezos_protocol_environment_009_PsFLoren__Environment.Fitness.t -> Tezos_protocol_environment_009_PsFLoren__Environment.Context.t -> t Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

Retrieves the state of the database and gives its abstract view. It also returns wether this is the first block validated with this version of the protocol.

type previous_protocol =

| Genesis of Tezos_raw_protocol_009_PsFLoren.Parameters_repr.t

| Edo_008

val prepare_first_block : level:int32 -> timestamp:Tezos_protocol_environment_009_PsFLoren__Environment.Time.t -> fitness:Tezos_protocol_environment_009_PsFLoren__Environment.Fitness.t -> Tezos_protocol_environment_009_PsFLoren__Environment.Context.t -> (previous_protocol * t) Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

val activate : t -> Tezos_protocol_environment_009_PsFLoren__Environment.Protocol_hash.t -> t Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

val recover : t -> Tezos_protocol_environment_009_PsFLoren__Environment.Context.t

Returns the state of the database resulting of operations on its abstract view

val current_level : t -> Tezos_raw_protocol_009_PsFLoren.Level_repr.t

val predecessor_timestamp : t -> Tezos_protocol_environment_009_PsFLoren__Environment.Time.t

val current_timestamp : t -> Tezos_protocol_environment_009_PsFLoren__Environment.Time.t

val current_fitness : t -> Tezos_protocol_environment_009_PsFLoren__Environment.Int64.t

val set_current_fitness : t -> Tezos_protocol_environment_009_PsFLoren__Environment.Int64.t -> t

val constants : t -> Tezos_raw_protocol_009_PsFLoren.Constants_repr.parametric

val patch_constants : t -> (Tezos_raw_protocol_009_PsFLoren.Constants_repr.parametric -> Tezos_raw_protocol_009_PsFLoren.Constants_repr.parametric) -> t Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

val first_level : t -> Tezos_raw_protocol_009_PsFLoren.Raw_level_repr.t

val add_fees : t -> Tezos_raw_protocol_009_PsFLoren.Tez_repr.t -> t Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult

Increment the current block fee stash that will be credited to baker's frozen_fees account at finalize_application

val add_rewards : t -> Tezos_raw_protocol_009_PsFLoren.Tez_repr.t -> t Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult

Increment the current block reward stash that will be credited to baker's frozen_fees account at finalize_application

val add_deposit : t -> Tezos_protocol_environment_009_PsFLoren__Environment.Signature.Public_key_hash.t -> Tezos_raw_protocol_009_PsFLoren.Tez_repr.t -> t Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult

Increment the current block deposit stash for a specific delegate. All the delegates' frozen_deposit accounts are credited at finalize_application

val get_fees : t -> Tezos_raw_protocol_009_PsFLoren.Tez_repr.t

val get_rewards : t -> Tezos_raw_protocol_009_PsFLoren.Tez_repr.t

val get_deposits : t -> Tezos_raw_protocol_009_PsFLoren.Tez_repr.t Tezos_protocol_environment_009_PsFLoren__Environment.Signature.Public_key_hash.Map.t

type Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.error += Gas_limit_too_high

val check_gas_limit : t -> 'a Tezos_raw_protocol_009_PsFLoren.Gas_limit_repr.Arith.t -> unit Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult

val set_gas_limit : t -> 'a Tezos_raw_protocol_009_PsFLoren.Gas_limit_repr.Arith.t -> t

val set_gas_unlimited : t -> t

val gas_level : t -> Tezos_raw_protocol_009_PsFLoren.Gas_limit_repr.t

val gas_consumed : since:t -> until:t -> Tezos_raw_protocol_009_PsFLoren.Gas_limit_repr.Arith.fp

val block_gas_level : t -> Tezos_raw_protocol_009_PsFLoren.Gas_limit_repr.Arith.fp

val init_storage_space_to_pay : t -> t

val update_storage_space_to_pay : t -> Tezos_protocol_environment_009_PsFLoren__Environment.Z.t -> t

val update_allocated_contracts_count : t -> t

val clear_storage_space_to_pay : t -> t * Tezos_protocol_environment_009_PsFLoren__Environment.Z.t * int

type Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.error += Undefined_operation_nonce

val init_origination_nonce : t -> Tezos_protocol_environment_009_PsFLoren__Environment.Operation_hash.t -> t

val origination_nonce : t -> Tezos_raw_protocol_009_PsFLoren.Contract_repr.origination_nonce Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult

val increment_origination_nonce : t -> (t * Tezos_raw_protocol_009_PsFLoren.Contract_repr.origination_nonce) Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult

val unset_origination_nonce : t -> t

Generic accessors

type key = string list

type value = bytes

type tree

module type T = Tezos_raw_protocol_009_PsFLoren.Raw_context_intf.T with type root := root and type key := key and type value := value and type tree := tree

include T with type T.t := t

type root

The type for root contexts.

include Tezos_raw_protocol_009_PsFLoren.Raw_context_intf.VIEW

type t

The type for context handler.

type tree

The type for context trees.

type key = string list

The type for context keys.

type value = bytes

The type for context values.

Getters

val mem : t -> key -> bool Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

mem t k is an Lwt promise that resolves to true iff k is bound to a value in t.

val mem_tree : t -> key -> bool Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

mem_tree t k is like mem but for trees.

val get : t -> key -> value Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

get t k is an Lwt promise that resolves to Ok v if k is bound to the value v in t and Storage_Error Missing_key otherwise.

val get_tree : t -> key -> tree Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

get_tree is like get but for trees.

val find : t -> key -> value option Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

find t k is an Lwt promise that resolves to Some v if k is bound to the value v in t and None otherwise.

val find_tree : t -> key -> tree option Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

find_tree t k is like find but for trees.

val list : t -> ?⁠offset:int -> ?⁠length:int -> key -> (string * tree) list Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

list t key is the list of files and sub-nodes stored under k in t. The result order is not specified but is stable.

offset and length are used for pagination.

Setters

val init : t -> key -> value -> t Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

init t k v is an Lwt promise that resolves to Ok c if:

k is unbound in t;

k is bound to v in c;

and c is similar to t otherwise.

It is Storage_error Existing_key if k is already bound in t.

val init_tree : t -> key -> tree -> t Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

init_tree is like init but for trees.

val update : t -> key -> value -> t Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

update t k v is an Lwt promise that resolves to Ok c if:

k is bound in t;

k is bound to v in c;

and c is similar to t otherwise.

It is Storage_error Missing_key if k is not already bound in t.

val update_tree : t -> key -> tree -> t Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

update_tree is like update but for trees.

val add : t -> key -> value -> t Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

add t k v is an Lwt promise that resolves to c such that:

k is bound to v in c;

and c is similar to t otherwise.

If k was already bound in t to a value that is physically equal to v, the result of the function is a promise that resolves to t. Otherwise, the previous binding of k in t disappears.

val add_tree : t -> key -> tree -> t Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

add_tree is like add but for trees.

val remove : t -> key -> t Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

remove t k v is an Lwt promise that resolves to c such that:

k is unbound in c;

and c is similar to t otherwise.

val remove_existing : t -> key -> t Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

remove_existing t k v is an Lwt promise that resolves to Ok c if:

k is bound in t to a value;

k is unbound in c;

and c is similar to t otherwise.

val remove_existing_tree : t -> key -> t Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

remove_existing_tree t k v is an Lwt promise that reolves to Ok c if:

k is bound in t to a tree;

k is unbound in c;

and c is similar to t otherwise.

val add_or_remove : t -> key -> value option -> t Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

add_or_remove t k v is:

add t k x if v is Some x;

remove t k otherwise.

val add_or_remove_tree : t -> key -> tree option -> t Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

add_or_remove_tree t k v is:

add_tree t k x if v is Some x;

remove t k otherwise.

Folds

val fold : ?⁠depth:[ `Eq of int | `Le of int | `Lt of int | `Ge of int | `Gt of int ] -> t -> key -> init:'a -> f:(key -> tree -> 'a -> 'a Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t) -> 'a Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

fold ?depth t root ~init ~f recursively folds over the trees and values of t. The f callbacks are called with a key relative to root. f is never called with an empty key for values; i.e., folding over a value is a no-op.

Elements are traversed in lexical order of keys.

The depth is 0-indexed. If depth is set (by default it is not), then f is only called when the conditions described by the parameter is true:

Eq d folds over nodes and contents of depth exactly d.

Lt d folds over nodes and contents of depth strictly less than d.

Le d folds over nodes and contents of depth less than or equal to d.

Gt d folds over nodes and contents of depth strictly more than d.

Ge d folds over nodes and contents of depth more than or equal to d.

module Tree : Tezos_raw_protocol_009_PsFLoren.Raw_context_intf.TREE with type t := t and type key := key and type value := value and type tree := tree

val project : t -> root

Internally used in Storage_functors to escape from a view.

val absolute_key : t -> key -> key

Internally used in Storage_functors to retrieve a full key from partial key relative a view.

type Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.error += Block_quota_exceeded

Raised if block gas quota is exhausted during gas consumption.

type Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.error += Operation_quota_exceeded

Raised if operation gas quota is exhausted during gas consumption.

val consume_gas : t -> Tezos_raw_protocol_009_PsFLoren.Gas_limit_repr.cost -> t Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult

Internally used in Storage_functors to consume gas from within a view. May raise Block_quota_exceeded or Operation_quota_exceeded.

val check_enough_gas : t -> Tezos_raw_protocol_009_PsFLoren.Gas_limit_repr.cost -> unit Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult

Check if consume_gas will fail

val description : t Tezos_raw_protocol_009_PsFLoren.Storage_description.t

val reset_internal_nonce : t -> t

Initialize the local nonce used for preventing a script to duplicate an internal operation to replay it.

val fresh_internal_nonce : t -> (t * int) Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult

Increments the internal operation nonce.

val record_internal_nonce : t -> int -> t

Mark an internal operation nonce as taken.

val internal_nonce_already_recorded : t -> int -> bool

Check is the internal operation nonce has been taken.

val allowed_endorsements : t -> (Tezos_protocol_environment_009_PsFLoren__Environment.Signature.Public_key.t * int list * bool) Tezos_protocol_environment_009_PsFLoren__Environment.Signature.Public_key_hash.Map.t

Returns a map where to each endorser's pkh is associated the list of its endorsing slots (in increasing order) for a given level.

val included_endorsements : t -> int

Keep track of the number of endorsements that are included in a block

val init_endorsements : t -> (Tezos_protocol_environment_009_PsFLoren__Environment.Signature.Public_key.t * int list * bool) Tezos_protocol_environment_009_PsFLoren__Environment.Signature.Public_key_hash.Map.t -> t

Initializes the map of allowed endorsements, this function must only be called once.

val record_endorsement : t -> Tezos_protocol_environment_009_PsFLoren__Environment.Signature.Public_key_hash.t -> t

Marks an endorsement in the map as used.

val fold_map_temporary_lazy_storage_ids : t -> (Tezos_raw_protocol_009_PsFLoren.Lazy_storage_kind.Temp_ids.t -> Tezos_raw_protocol_009_PsFLoren.Lazy_storage_kind.Temp_ids.t * 'res) -> t * 'res

val map_temporary_lazy_storage_ids_s : t -> (Tezos_raw_protocol_009_PsFLoren.Lazy_storage_kind.Temp_ids.t -> (t * Tezos_raw_protocol_009_PsFLoren.Lazy_storage_kind.Temp_ids.t) Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t) -> t Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t