Module Sapling.Roots

type t

type context = t

type key = int32

An abstract type for keys

type value = Tezos_protocol_environment_009_PsFLoren__Environment.Sapling.Hash.t

The type of values

val mem : context -> key -> bool Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

Tells if a given key is already bound to a storage bucket

val get : context -> key -> value Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

Retrieve a value from the storage bucket at a given key ; returns Storage_error Missing_key if the key is not set ; returns Storage_error Corrupted_data if the deserialisation fails.

val find : context -> key -> value option Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

Retrieve a value from the storage bucket at a given key ; returns None if the value is not set ; returns Storage_error Corrupted_data if the deserialisation fails.

val update : context -> key -> value -> Tezos_raw_protocol_009_PsFLoren.Raw_context.t Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

Updates the content of a bucket ; returns A Storage_Error Missing_key if the value does not exists.

val init : context -> key -> value -> Tezos_raw_protocol_009_PsFLoren.Raw_context.t Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

Allocates a storage bucket at the given key and initializes it ; returns a Storage_error Existing_key if the bucket exists.

val add : context -> key -> value -> Tezos_raw_protocol_009_PsFLoren.Raw_context.t Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

Allocates a storage bucket at the given key and initializes it with a value ; just updates it if the bucket exists.

val add_or_remove : context -> key -> value option -> Tezos_raw_protocol_009_PsFLoren.Raw_context.t Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

When the value is Some v, allocates the data and initializes it with v ; just updates it if the bucket exists. When the value is None, delete the storage bucket when the value ; does nothing if the bucket does not exists.

val remove_existing : context -> key -> Tezos_raw_protocol_009_PsFLoren.Raw_context.t Tezos_protocol_environment_009_PsFLoren__Environment.Error_monad.tzresult Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

Delete a storage bucket and its contents ; returns a Storage_error Missing_key if the bucket does not exists.

val remove : context -> key -> Tezos_raw_protocol_009_PsFLoren.Raw_context.t Tezos_protocol_environment_009_PsFLoren__Environment.Lwt.t

Removes a storage bucket and its contents ; does nothing if the bucket does not exists.